requirement.txt
greenlet==2.0.1
mysql-connector-python==8.0.31
numpy==1.22.4
pandas==1.4.2
protobuf==3.20.1
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.7
six==1.16.0
SQLAlchemy==1.4.37
